refactor(types): address PR review feedback
- Move RoomConnection to types/room.ts
- Refine BeneficiarySummary status type in types/models.ts
- Cleanup types/admin.ts verification
